associate user to post and comment and authorize users

+++ blog_on_rails/app/controllers/comments_controller.rb
@@ -5,18 +5,28 @@ class CommentsController < ApplicationController
         @comment = Comment.new comment_params
         @post = Post.find params[:post_id]
         @comment.post = @post
-
-        if @comment.save
-            redirect_to post_path(@post), notice: "New Comment!"
+        @comment.user = current_user
+        if can? :crud, @comment
+            if @comment.save
+                redirect_to post_path(@post), notice: "New Comment!"
+            else
+                @comments = @post.comments.order(created_at: :desc)
+                render 'posts/show', alert: "error!"
+            end
         else
-            @comments = @post.comments.order(created_at: :desc)
-            render 'posts/show', alert: "error!"
+            flash[:alert] = 'Not Authorized'
+            redirect_to root_path
         end
     end
 
     def destroy
         comment = Comment.find params[:id]
-        comment.destroy
-        redirect_to post_path(comment.post_id)
+        if can? :crud, comment
+            comment.destroy
+            redirect_to post_path(comment.post_id)
+        else
+            flash[:alert] = 'Not Authorized'
+            redirect_to root_path
+        end
     end
 end
